Escape URL when writing to log
The request URL is written without escaping to the log file. This
could be an issue, because a malicious attacker can use this to forge
log entries. It can also be used to generate invalid log files which
could corrupt log parsing tools.

The proposed solution is to escape using percent encoding the URL,
leaving safe characters unescaped.

A sample exploit:

  bash$ curl localhost:3000/'%20from%208.8.8.8%0a%5b12/34/56%2069:96:42%20PM%5d%20GETOUTTAHERE%20Hello%20world!'